Fix bug with sum of CartTotal

+++ Mango.Services.EmailAPI/Models/Dto/CartHeaderDto.cs
@@ -9,7 +9,7 @@ namespace Mango.Services.EmailAPI.Models.Dto
         public string? UserId { get; set; }
         public string? CouponCode { get; set; }
         public double Discount { get; set; }
-        public double CartTotal { get; set; }
+        public decimal CartTotal { get; set; }
         public string? Name { get; set; }
         public string? Phone { get; set; }
         public string? Email { get; set; }
