+++ Mango.Services.ShoppingCart/Models/Dto/CartHeaderDto.cs
@@ -6,7 +6,7 @@
         public string? UserId { get; set; }
         public string? CouponCode { get; set; }
         public double Discount { get; set; }
-        public double CartTotal { get; set; }
+        public decimal CartTotal { get; set; }
         public string? Name { get; set; }
         public string? Phone { get; set; }
         public string? Email { get; set; }
